Set dunst to print notifications on primary screen
The primary screen can be set by xrandr so I can set dunst to print
notifications always on monitor 0 (the first monitor = the primary
screen) and set primary screen properly by xrandr.  This kind of logic
does not require any dunst restart due to change of the monitor and it
works like charm.  The docker-mode-toggle script has been changed
accordingly in order to set primary screen on the largest monitor
connected to the docking station.