Set proper screen layout for LightDM
I found out that it's possible to set proper screen layout for LightDM
before showing the login screen, so the login dialog, that is always
being shown on the primary monitor will be displayed on external screen.
When laptop is attached to the docking station this will also set the
base for the xrandr settings being used when i3 is launched.  For this
purpose I defined set-screen-layout script that is being executed by
LightDM.  Since this script will be maintained together with the
docked-mode-toggle, it's conveinient to keep it in the ~/.i3/scripts
directory.